==> 1_Sunspot_Data.ipynb <==
#Code notebook for converting data into csv format for sunspots 
import pandas as pd
import numpy as np

df = pd.read_csv('/content/usaf_solar-region-reports_2016.txt', header = None)

row_list = []
for i in range(len(df)):
  a = df.iloc[i,0]
  dict1 = {}
  word = [char for char in a]

  Code = {'Code' : ''.join(word[0:2])}
  dict1.update(Code)

  Year = {'Year' : ''.join(word[2:4])}
  dict1.update(Year)

  Month = {'Month' : ''.join(word[4:6])}
  dict1.update(Month)

  Day = {'Day' : ''.join(word[6:8])}
  dict1.update(Day)

  Hours = {'Hours' : ''.join(word[9:11])}
  dict1.update(Hours)

  Minutes = {'Minutes' : ''.join(word[11:13])}
  dict1.update(Minutes)

  Latitude = {'Latitude' : ''.join(word[14:18])}
  dict1.update(Latitude)

  Meridian_Distance = {'Meridian Distance' : ''.join(word[18:20])}
  dict1.update(Meridian_Distance)

  Wilson_Magnetic_Classification = {'Wilson Magnetic Classification' : ''.join(word[20:24])}
  dict1.update(Wilson_Magnetic_Classification)

  Max_Magnetic_Field_Strength = {'Max Magnetic Field Strength' : ''.join(word[25:26])}
  dict1.update(Max_Magnetic_Field_Strength)

  Mt_Wilson_Spot_GN = {'Mt. Wilson Spot GN (MWIL)' : ''.join(word[27:32])}
  dict1.update(Mt_Wilson_Spot_GN)

  MWIL_Subscript = {'MWIL Subscript' : ''.join(word[32:33])}
  dict1.update(MWIL_Subscript)

  NOAA_Number = {'NOAA Number' : ''.join(word[33:38])}
  dict1.update(NOAA_Number)

  NOAA_Number_sbscript = {'NOAA Number sbscript' : ''.join(word[38:39])}
  dict1.update(NOAA_Number_sbscript)

  Zurich_Class = {'Zurich Class' : ''.join(word[39:40])}
  dict1.update(Zurich_Class)

  penumbra_Class = {'penumbra Class' : ''.join(word[40:41])}
  dict1.update(penumbra_Class)

  Compactness_Class = {'Compactness Class' : ''.join(word[41:42])}
  dict1.update(Compactness_Class)

  Number_of_Spots = {'Number of Spots' : ''.join(word[43:45])}
  dict1.update(Number_of_Spots)

  Longitudanal_extent = {'Longitudanal extent' : ''.join(word[46:48])}
  dict1.update(Longitudanal_extent)

  Area = {'Area' : ''.join(word[48:52])}
  dict1.update(Area)

  Individual_Central_Meridian_Passage_year = {'Individual Central Meridian Passage year' : ''.join(word[53:55])}
  dict1.update(Individual_Central_Meridian_Passage_year)

  Individual_Central_Meridian_Passage_month = {'Individual Central Meridian Passage month' : ''.join(word[55:57])}
  dict1.update(Individual_Central_Meridian_Passage_month)

  Individual_Central_Meridian_Passage_day = {'Individual Central Meridian Passage day' : ''.join(word[57:61])}
  dict1.update(Individual_Central_Meridian_Passage_day)

  Regional_Central_Meridian_Passage_year = {'Regional Central Meridian Passage year' : ''.join(word[62:64])}
  dict1.update(Regional_Central_Meridian_Passage_year)

  Regional_Central_Meridian_Passage_month = {'Regional Central Meridian Passage month' : ''.join(word[64:66])}
  dict1.update(Regional_Central_Meridian_Passage_month)

  Regional_Central_Meridian_Passage_day = {'Regional Central Meridian Passage day' : ''.join(word[66:70])}
  dict1.update(Regional_Central_Meridian_Passage_day)

  Station_serial_number = {'Station serial number' : ''.join(word[71:74])}
  dict1.update(Station_serial_number)

  Quality = {'Quality' : ''.join(word[75:76])}
  dict1.update(Quality)

  Station = {'Station' : ''.join(word[76:80])}
  dict1.update(Station)



  row_list.a(dict1)
      
      

sunspot_1996 = pd.DataFrame(row_list) 

sunspot_1997 = pd.DataFrame(row_list) 

sunspot_1998 = pd.DataFrame(row_list) 

sunspot_1999 = pd.DataFrame(row_list) 

sunspot_2000 = pd.DataFrame(row_list) 

sunspot_2001 = pd.DataFrame(row_list) 

sunspot_2002 = pd.DataFrame(row_list) 

sunspot_2003 = pd.DataFrame(row_list) 

sunspot_2004 = pd.DataFrame(row_list) 

sunspot_2005 = pd.DataFrame(row_list) 

sunspot_2006 = pd.DataFrame(row_list) 

sunspot_2007 = pd.DataFrame(row_list) 

sunspot_2008 = pd.DataFrame(row_list) 

sunspot_2009 = pd.DataFrame(row_list) 

sunspot_2010 = pd.DataFrame(row_list) 

sunspot_2011 = pd.DataFrame(row_list) 

sunspot_2012 = pd.DataFrame(row_list) 

sunspot_2013 = pd.DataFrame(row_list) 

sunspot_2014 = pd.DataFrame(row_list) 

sunspot_2015 = pd.DataFrame(row_list) 

sunspot_2016 = pd.DataFrame(row_list) 

#sunspot_2000['Year'] = 2000
#sunspot_2001['Year'] = 2001
#sunspot_2002['Year'] = 2002
#sunspot_2003['Year'] = 2003
#sunspot_2004['Year'] = 2004
#sunspot_2005['Year'] = 2005
#sunspot_2006['Year'] = 2006
#sunspot_2007['Year'] = 2007
#sunspot_2008['Year'] = 2008
#sunspot_2009['Year'] = 2009
#sunspot_2010['Year'] = 2010
#sunspot_2011['Year'] = 2011
#sunspot_2012['Year'] = 2012
#sunspot_2013['Year'] = 2013
#sunspot_2014['Year'] = 2014
#sunspot_2015['Year'] = 2015
#sunspot_2016['Year'] = 2016

frames2 = [sunspot_1996, sunspot_1997, sunspot_1998, sunspot_1999]
sunspot_1990 = pd.concat(frames2)

sunspot_1990

sunspot_1990.drop(columns = ['Max Magnetic Field Strength', 'Mt. Wilson Spot GN (MWIL)', 'MWIL Subscript', 'NOAA Number sbscript', 'Individual Central Meridian Passage year',
                                'Individual Central Meridian Passage month', 'Individual Central Meridian Passage day',
                               'Regional Central Meridian Passage year',
       'Regional Central Meridian Passage month',
       'Regional Central Meridian Passage day' ], inplace = True)

sunspot_1990.drop(columns = ['Code', 'Latitude', 'Meridian Distance', 'Station serial number', 'Quality', 'Station'], inplace = True)

#sunspot_1990.drop(columns = ['Hours', 'Minutes'], inplace = True)

sunspot_1990= sunspot_1990.replace('  ', np.nan)

sunspot_1990.isna().sum()

sunspot_1990.dropna(inplace = True)

#sunspot_1990['Year'] = sunspot_1990['Year'].astype('int') + 1900
#sunspot_1990['Date'] = sunspot_1990["Year"].astype('str') +  sunspot_1990["Month"].astype('str') + sunspot_1990["Day"].astype('str') + sunspot_1990['Hours'] + sunspot_1990['Minutes']

#sunspot_1990.drop(columns = ['Year', 'Month', 'Day', 'Hours', 'Minutes'], inplace = True)
sunspot_1990['Date'] = pd.to_datetime(sunspot_1990['Date'], format='%Y%m%d%H%M%S')

sunspot_1990

frames = [sunspot_2000, sunspot_2001, sunspot_2002, sunspot_2003, sunspot_2004,
          sunspot_2005, sunspot_2006, sunspot_2007, sunspot_2008,
          sunspot_2009, sunspot_2010, sunspot_2011, sunspot_2012, sunspot_2013, sunspot_2014, sunspot_2015, sunspot_2016]

sunspot_regions = pd.concat(frames)

sunspot_regions

sunspot_regions= sunspot_regions.replace('', np.nan)

sunspot_regions.isna().sum()

sunspot_regions.dropna(inplace= True)

sunspot_regions2 = sunspot_regions.copy()

sunspot_regions.drop(columns = ['Max Magnetic Field Strength', 'Mt. Wilson Spot GN (MWIL)', 'MWIL Subscript', 'NOAA Number sbscript', 'Individual Central Meridian Passage year',
                                'Individual Central Meridian Passage month', 'Individual Central Meridian Passage day',
                               'Regional Central Meridian Passage year',
       'Regional Central Meridian Passage month',
       'Regional Central Meridian Passage day', 'Code', 'Station serial number', 'Quality',
       'Station', 'Latitude', 'Meridian Distance' ], inplace = True)

sunspot_regions['Year'] = sunspot_regions['Year'].astype('int') +2000
sunspot_regions['Date'] = sunspot_regions["Year"].astype('str') +  sunspot_regions["Month"].astype('str') + sunspot_regions["Day"].astype('str') + sunspot_regions['Hours'] + sunspot_regions['Minutes']

sunspot_regions.drop(columns = ['Year', 'Month', 'Day', 'Hours', 'Minutes'], inplace = True)
sunspot_regions['Date'] = pd.to_datetime(sunspot_regions['Date'], format='%Y%m%d%H%M%S')

#sunspot_regions['Latitude_Longitude'] =  sunspot_regions['Latitude'] + sunspot_regions['Meridian Distance']
#sunspot_regions.drop(columns = ['Latitude', 'Meridian Distance' ], inplace = True)

#sunspot_regions['McIntosh_Classification'] = sunspot_regions['Zurich Class'] + sunspot_regions['penumbra Class'] + sunspot_regions['Compactness Class']
#sunspot_regions.drop(columns = ['Zurich Class', 'penumbra Class', 'Compactness Class' ], inplace = True)

#sunspot_regions['Time'] = sunspot_regions['Hours'] + ':' + sunspot_regions['Minutes']
#sunspot_regions.drop(columns = ['Hours', 'Minutes'], inplace= True)

sunspot_regions

sunspot_regions = sunspot_regions.replace(' ////', np.nan)

sunspot_regions.isna().sum()

sunspot_regions.dropna(inplace = True)

sunspot_regions

frames = [sunspot_1990, sunspot_regions]
total = pd.concat(frames)

total.to_csv('sunspot_regions.csv')

total.info()

total = total.replace('           ', np.nan)

total.isna().sum()

total.to_excel('sunspots.xlsx')

total
