==> 3_Associated_Unassociated_spots.ipynb <==
import pandas as pd 
import numpy as np 
import seaborn as sns 
import matplotlib.pyplot as plt 
%matplotlib inline

flares = pd.read_excel('C:/Users/Raghav/Desktop/Solar Flare/Data/Solarflares.xlsx')
flares = flares.rename(columns = {'NOAA_Number' : 'NOAA_Number_flare'})
flares['Date'] = pd.to_datetime(flares['Date'])
flares['NOAA_Number_flare'] = flares['NOAA_Number_flare'].replace('     ', np.nan)
flares['NOAA_Number_flare'] = flares['NOAA_Number_flare'].replace('    ', np.nan)
flares.info()

flares = flares[flares['X_Ray_Class'].isin(['X', 'M', 'C'])]
flares = flares[flares['NOAA_Number_flare'].notnull()]
flares['NOAA_Number_flare'] = flares['NOAA_Number_flare'].astype('int')

spots = pd.read_excel('C:/Users/Raghav/Desktop/Solar Flare/Data/sunspots.xlsx')
spots = spots.rename(columns = {'NOAA Number' : 'NOAA_Number_spots'})
spots['NOAA_Number_spots'] = spots['NOAA_Number_spots'].replace('          ', np.nan)
spots.dropna(inplace= True)
spots['Date'] = pd.to_datetime(spots['Date'])
spots['NOAA_Number_spots'] = spots['NOAA_Number_spots'].astype('object') 
spots.info()

spots = spots.replace('     ' , np.nan)
spots = spots.dropna()
spots['NOAA_Number_spots'] = spots['NOAA_Number_spots'].astype('int')
spots.info()

import datetime
from tqdm import tqdm

flare = []
for index,row in tqdm(flares.iterrows(),total=flares.shape[0]):
    noa_flare = row['NOAA_Number_flare']
    x_ray = row['X_Ray_Class']
    flare_date = row['Date']
    d1 ={}
    date_flare = row['Date']
    if x_ray == 'X' or x_ray == 'M' or x_ray == 'C':
      if noa_flare != np.nan:
        for i,r in spots.iterrows():
          noa_spot = r['NOAA_Number_spots']
          spot_date = r['Date']
          if noa_flare == noa_spot:
            diff_time = (flare_date - spot_date)/60
            a = []
            if diff_time < datetime.timedelta(minutes = 480):
              a = pd.concat([row,r],axis=0)
              a = a.to_dict()
              a["Status"] = "Associated"
              flare.append(a)
            break

associated_spots = pd.DataFrame(flare)

associated_spots

#Create an outer join for the unassociated data points
a = (pd.merge(spots,flares, indicator=True, how='outer')
         .query('_merge=="left_only"')
         .drop('_merge', axis=1))

a

b= a.sort_values('Date', ascending=False).drop_duplicates(subset=['NOAA_Number_spots'])

sunspots_target = b.sort_values('Date')

sunspots_target = sunspots_target.reset_index()

sunspots_target['NOAA_Number_spots'] = sunspots_target['NOAA_Number_spots'].astype('int')

sunspots_target['Status'] = 'Unassociated'

sunspots_target

associated_spots['X_Ray_Class'].value_counts()

associated_spots.to_excel("associated_spots_new.xlsx")

sunspots_target.to_excel('unassociated_spots_new.xlsx')
