==> 4_Combining_data.ipynb <==
import pandas as pd
import numpy as np 
import matplotlib.pyplot as plt
import seaborn as sns 
%matplotlib inline

unassociated = pd.read_excel(r'C:\Users\Raghav\Desktop\Solar Flare\Data\Unassociated_spots.xlsx')

#Contains 4425 rows of data - For spots that did not produce falres 
unassociated

#Drop columns that are not required
unassociated = unassociated.drop(columns =['Unnamed: 0', 'level_0', 'index', 'Integrated_Flux', 'NOAA_Number_flare'], axis= 1)

associated = pd.read_excel(r'C:\Users\Raghav\Desktop\Solar Flare\Data\associated_spots_new.xlsx')

#Contains 13664 rows fo data for spots that prdocued flares 
associated

#drop columns that are not required
associated.drop(columns = ['Unnamed: 0', 'Integrated_Flux', 'NOAA_Number_flare'], axis = 1, inplace = True)

associated_C = associated[associated['X_Ray_Class'] == 'C']

associated_C

associated_MX = associated[associated['X_Ray_Class'] != 'C']

associated_MX

# Randomly sample spots that rpdouced flares of C class to balance datasets 
associated_C_randpm = associated_C.sample(n = 2652)

associated_C_randpm

associated_4225 = [associated_MX, associated_C_randpm]
associated_new = pd.concat(associated_4225)

#Contains 4425 rows for spots that produced flares 
associated_new

flares = [associated_new, unassociated]

Flares_Data = pd.concat(flares)

Flares_Data

Flares_Data[Flares_Data['Status'] == 'Associated']

#Rename columns
Flares_Data.rename(columns= {'Wilson Magnetic Classification' : 'magnetic_classification',
                             'NOAA_Number_spots': 'NOAA',
                             'Zurich Class' : 'zurich_class',
                             'penumbra Class' : 'penumbra_class',
                             'Compactness Class' : 'compactness_class',
                             'Number of Spots' : 'No_of_spots',
                             'Longitudanal extent' :'longitudanal_extent'}, inplace = True)
Flares_Data

Flares_Data.to_excel('sunspots_new.xlsx')

Flares_Data['Status'].value_counts()
