==> 5_Cleaning_Final_Data.ipynb <==
import pandas as pd 
import numpy as np 
import seaborn as sns 
import matplotlib.pyplot as plt 
%matplotlib inline

df = pd.read_excel(r'C:\Users\Raghav\Desktop\Solar Flare\Data\Final Data\sunspots_new.xlsx')

df.info()

df = df.replace('           ', np.nan)
df = df.replace('', np.nan)
df = df.replace(' ', np.nan)
df = df.replace('  ', np.nan)
df = df.replace('   ', np.nan)
df = df.replace('    ', np.nan)
df = df.replace('     ', np.nan)

df.isna().sum()

df = df.drop(columns = ['X_Ray_Class', 'X_Ray_Intensity', 'Date', 'NOAA'])

df = df.drop(columns = ['Unnamed: 0'])

df = df.dropna()

df.isna().sum()

df.head()

df.tail()

df['No_of_spots'] = df['No_of_spots'].astype('int')
df['longitudanal_extent'] = df['longitudanal_extent'].astype('int')
df['Area'] = df['Area'].astype('int')
#, 'longitudanal_extent', 'Area']

df['Status'].value_counts()

df_majority = df[df['Status'] == 'Associated']
df_minority = df[df['Status'] == 'Unassociated']

from sklearn.utils import resample
df_upsampled = resample(df_minority,
                        replace = True,
                        n_samples = 3784)

df_upsampled['Status'].value_counts()

sunspots = pd.concat([df_upsampled, df_majority])

sunspots = sunspots.reset_index()

sunspots.head()

sunspots.drop(columns= ['index'], inplace= True )

sunspots.to_excel('sunspots_cleaned.xlsx')
